Endpoint of Update finished

+++ crud_basic/src/main/java/com/sena/crud_basic/service/GenericService.java
@@ -3,6 +3,8 @@ package com.sena.crud_basic.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import jakarta.persistence.EntityNotFoundException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -17,8 +19,17 @@ public abstract class GenericService<T, D> {
     // Método abstracto para conversión de Modelo a DTO
     protected abstract D convertToDTO(T model);
 
-    // Guardar o actualizar
+    // Guardar 
     public T save(D dto) {
+        T entity = convertToModel(dto);
+        return repository.save(entity);
+    }
+
+     // Actualizar un registro existente
+    public T update(Integer id, D dto) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Entidad con ID " + id + " no encontrada.");
+        }
         T entity = convertToModel(dto);
         return repository.save(entity);
     }
@@ -30,11 +41,17 @@ public abstract class GenericService<T, D> {
 
     // Buscar por ID
     public Optional<T> findById(Integer id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Entidad con ID " + id + " no encontrada.");
+        }
         return repository.findById(id);
     }
 
     // Eliminar por ID
     public void delete(Integer id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Entidad con ID " + id + " no encontrada.");
+        }
         repository.deleteById(id);
     }
 }
